requirements.txt
mcp
google-genai
psycopg[binary]
python-dotenv
streamlit
pandas
openpyxl
